♻️ refactor : apps/users/utils 폴더 생성 후 리팩토링
유저 앱에서 사용되는 유틸 기능들을 정리
