+++ apps/users/utils/jwt.py
@@ -3,7 +3,7 @@ from typing import Dict
 
 from rest_framework_simplejwt.tokens import RefreshToken
 
-from .models import User
+from apps.users.models import User
 
 
 class JWTService:
