+++ apps/users/views/kakao_view.py
@@ -2,9 +2,9 @@ from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from ..jwt import JWTService
-from ..serializers import KakaoLoginSerializer, LoginResponseSerializer, UserSerializer
-from ..social_auth import SocialAuthService
+from apps.users.utils.jwt import JWTService
+from ..serializers import KakaoLoginSerializer, UserSerializer
+from apps.users.utils.social_auth import SocialAuthService
 from ..social_login.kakao_service import KakaoService
